+++ NEWS.md
@@ -1,4 +1,4 @@
-# emuR 0.1.9.9000
+# emuR 0.1.9.9001
 
 ## bug fixes
 
@@ -8,6 +8,7 @@
 
 * added verbose parameter to export\_TextGridCollection()
 * improved pre-check of dir exists in export\_TextGridCollection()
+* added new replace\_itemLabels function
 
 # emuR 0.1.9